[Spring 지하철 노선도 - 3단계] 후니(최재훈) 미션 제출합니다. (#263)
* docs : 기능 구현 요구사항 추가

* feat: 동일한 이름 등록시 에러가 발생한다.

* refactor : domain이 nullable하지 않게 수정

* refactor : test 필드 접근제어자 추가

* refactor : 메서드명같은 변수명 수정

* refactor : 컨트롤러 불필요한 중복 제거

* refactor : 직렬화를 위한 빈 생성자 private 선언

* feat : 데이터 접근 에러에 대한 핸들링 추가

* feat : section 도메인 생성

* feat : 이미 존재하는 구간 등록 시 예외 발생 기능 추가

* feat : 연결되지 않는 섹션은 추가할 수 없는 기능 추가

* feat : 구간 연결될 시 예외 발생 기능 추가

* docs : 기능 구현 목록 체크

* feat : 구간 중간 추가 기능 추가

* feat : 업데이트 한 구간 찾기 기능 추가

* docs : 구간삭제 구현 목록 추가

* feat : 구간 삭제 기능 추가

* test : 구간이 하나일 경우 삭제 불가 테스트 추가

* faet : 섹션끼리 합치는 기능 추가

* test : 삭제 요청 테스트 메서드 분리

* test : acceptance test setup 변경

* refactor : equals 재정의를 id로 설정

* refactor : section, station equals 수정

* refactor : 테스트에 전달되는 request 변경

* refactor : Line create 기능 변경

* test : SectionRepository test 추가

* feat : 구간 생성 기능 추가

* feat : 구간 삭제 기능 추가

* feat : 첫번째 역 찾기 기능 추가

* refactor : 노선 찾아오는 쿼리 수정

* refactor : line delete 로직 수정

* refactor : 인수테스트 사례 추가

* refactor : delete 실패 커스텀 예외 추가

* test : 리팩터링으로 인해 수정해야 하는 테스트 변경

* feat : null 입력에 대한 handler 추가

* test : update 테스트 추가

* test : 여러 구간 삭제 테스트 추가

* fix : line 가져오기 오타 수정

* fix : update시 거리가 수정 안되는 버그 해결

* refactor : 사용하지 않는 코드 제거

* test : 반환하는 값에 대해 순서와 값을 보장하는지 테스트

* test : 사용하지 않는 테스트 삭제

* test : sectionService 테스트 추가

* test : 동일한 로직이 반복되는 테스트를 nested로 분리

* refactor : null체크를 가장 위에서 처리

* test : private 키워드 추가

* refactor : 패키지 명 repository로 변경

* refactor : sql 예약어 대문자로 변경

* refactor : section을 찾을때 파라미터를 한개로 변경

* refactor : optional 파라미터 제거 및 이름 예외처리 수정

* refactor : null대신 빈 리스트를 가진 일급 컬렉션 반환

* refactor : 코드 스타일 수정

* refactor : LineSection 객체를 분리

* refactor : 객체 생성시 null 방지

* refactor : Controller advice 메서드 명 수정

* refactor : 에러 클래스 변경에 따른 ExceptionHandler 수정

* refactor : Transactional 애노테이션 추가

* refactor : sections update 로직 수정

* refactor : pickUpdate 로직 이중 포문 제거

* refactor : delete 로직 collection api 사용

* refactor : 정렬 로직 도메인으로 이동

* refactor : test fixture를 이용하여 변경

* refactor : 에러에 대한 내용 Logging

* feat : logback 설정 추가

* refactor : 불필요한 import 제거

* refactor : 전체 예외 핸들러에 대한 반환 메시지 변경

* refactor : Line 최종 생성으로 변경

* refactor : validate를 묶는 메서드 추가

* refactor : station이 구간에 존재할 경우 delete 실패

* refactor : station 이름 중복을 sql 쿼리로 확인

* refactor : line이름 중복을 sql 쿼리로 확인

* refactor : logback fileNamePattern 변경

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -1,39 +1,45 @@
 package wooteco.subway.service;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import wooteco.subway.dao.StationRepository;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
+import wooteco.subway.repository.SectionRepository;
+import wooteco.subway.repository.StationRepository;
 import wooteco.subway.utils.exception.NameDuplicatedException;
+import wooteco.subway.utils.exception.SubwayException;
 
+@Transactional
 @Service
 public class StationService {
 
     private final StationRepository stationRepository;
+    private final SectionRepository sectionRepository;
 
-    public StationService(StationRepository stationRepository) {
+    public StationService(StationRepository stationRepository,
+                          SectionRepository sectionRepository) {
         this.stationRepository = stationRepository;
+        this.sectionRepository = sectionRepository;
     }
 
-    @Transactional
     public StationResponse save(final StationRequest stationRequest) {
-        validateDuplicateName(stationRepository.findByName(stationRequest.getName()));
+        String name = stationRequest.getName();
+        validateDuplicateName(stationRepository.isNameExists(name), name);
 
-        Station station = stationRepository.save(new Station(stationRequest.getName()));
+        Station station = stationRepository.save(new Station(name));
         return new StationResponse(station.getId(), station.getName());
     }
 
-    private void validateDuplicateName(final Optional<Station> station) {
-        station.ifPresent(s -> {
-            throw new NameDuplicatedException(NameDuplicatedException.NAME_DUPLICATE_MESSAGE);
-        });
+    private void validateDuplicateName(final boolean isDuplicateName, final String name) {
+        if (isDuplicateName) {
+            throw new NameDuplicatedException(NameDuplicatedException.NAME_DUPLICATE_MESSAGE + name);
+        }
     }
 
+    @Transactional(readOnly = true)
     public List<StationResponse> showStations() {
         return stationRepository.findAll().stream()
                 .map(StationResponse::new)
@@ -41,6 +47,9 @@ public class StationService {
     }
 
     public void deleteStation(final Long id) {
+        if (sectionRepository.isStationExist(id)) {
+            throw new SubwayException("[ERROR] 역이 구간에 존재하여 삭제할 수 없습니다.");
+        };
         stationRepository.deleteById(id);
     }
 }
